requirements_dev.txt
-r Requirements/local.txt

django-debug-toolbar==4.3.0